replaced output service
